[OpenMP][CodeGen][Runtime] Introduce 'perfo' schedule
Introducing the 'perfo' schedule algorithm, designed to be utilized in
conjunction with the 'for' directive and the clause 'schedule'. The
'perfo' algorithm is part of a set of proposed approximate algorithms,
with a primary goal of optimizing code performance while preserving
programmer control. By selectively skipping iterations within a loop,
providing a flexible approach to boost overall efficiency.